playing with model spec and feature space.
Concepts are supposed to represent "views" into a data set, i.e. a
feature space describes a (somehow relevant) subset of the features in a
dataset. A model spec additionally contains the definition of a response
variable, and can be passed into the training procedure of a model
fitting alongside a dataset.