feature scaling: recover original coefficients.
Solving the linear equation after applying feature scaling produces
different coefficients. Since we may want to work with the original
values in the original scaling of the data, we need to copmute the
coefficients as they would have been computed in the original problem.

Note: unit test only, to work out and validate the formula.