scaling cost function: average over cases.
This is useful because the gradient otherwise depends on the sample
size, i.e. duplicating the training set leads to larger gradients (2x)
and you have reduce the learn rate to compensate.

Normalizing by dividing by the sample size helps to keep the numbers in
a value range irrespective of batch size / sample size.